requirements.txt
arrow==0.12.1
behave==1.2.6
behave-django==1.1.0
dj-database-url==0.5.0
Django==2.1.1
django-admin==1.3.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-excel-response2==2.0.8
django-extensions==2.1.2
django-six==1.0.4
factory-boy==2.11.1
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
screen==1.0.1
selenium==3.14.1
six==1.11.0
whitenoise==3.3.1
xlwt==1.3.0
